removed warning due to type conversion

+++ base/samples/frame.h
@@ -149,7 +149,7 @@ namespace base { namespace samples { namespace frame {
                     size = getPixelSize() * getPixelCount();
                 else
                     if(!isCompressed() && size != getPixelSize()*getPixelCount()){
-                        printf("Expect image size: %i, got: %i\n",getPixelSize()*getPixelCount(),size);
+                        printf("Expect image size: %i, got: %i\n",getPixelSize()*getPixelCount(),(int)size);
                         throw std::runtime_error("Frame::init: wrong image size!");
                     }
